+++ src/yafowil/widget/select2/resources/widget.js
@@ -21,11 +21,42 @@ if (typeof(window.yafowil) == "undefined") yafowil = {};
     });
 
     $.extend(yafowil, {
+
         select2: {
+
+            extract_value: function(value) {
+                if (value.indexOf('javascript:') == 0) {
+                    value = value.substring(11, value.length);
+                    value = value.split('.');
+                    if (!value.length) {
+                        throw "No function defined";
+                    }
+                    var ctx = window;
+                    var name;
+                    for (var idx in value) {
+                        name = value[idx];
+                        if (typeof(ctx[name]) == "undefined") {
+                            throw "'" + name + "' not found";
+                        }
+                        ctx = ctx[name];
+                    }
+                    value = ctx;
+                }
+                return value;
+            },
+
+            update_options: function(options) {
+                for (var idx in options) {
+                    name = options[idx];
+                    options[name] = select2.extract_value(options[name]);
+                }
+                return options;
+            },
+
             binder: function(context) {
                 $('.select2', context).each(function(event) {
                     var elem = $(this);
-                    var options = elem.data();
+                    var options = select2.update_options(elem.data());
                     elem.select2(options);
                 });
             }