+++ HISTORY.rst
@@ -4,6 +4,10 @@ History
 1.1.dev0
 --------
 
+- Remove ``select2input`` blueprint, use ``inputtag`` property of ``select2``
+  blueprint instead.
+  [rnix, 2014-06-19]
+
 - Provide all available constructor options in blueprint.
   [rnix, 2014-06-19]